requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
supabase==2.0.2
langchain-openai==0.0.2
python-dotenv==1.0.0
langchain==0.0.350